Added a helper file for SubmitBookingPopUp component.
